doc: 로또 클래스 main 패키지로 이동
