+++ src/test/kotlin/lotto/LottoNumbersTest.kt
@@ -5,26 +5,6 @@ import io.kotest.matchers.collections.shouldContainInOrder
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
 
-class Lotto(numbers: List<LottoNumber>) {
-    val lottoNumbers: List<LottoNumber> = numbers.map { it.copy() }.sortedBy { it.getNumber() }
-
-    init {
-        require(numbers.size == 6) { INIT_ERROR_MESSAGE }
-    }
-
-    fun getSize(): Int {
-        return lottoNumbers.size
-    }
-
-    fun getNumbers(): List<LottoNumber> {
-        return lottoNumbers
-    }
-
-    companion object {
-        private const val INIT_ERROR_MESSAGE: String = "로또 번호는 6개여야 합니다."
-    }
-}
-
 class LottoNumbersTest {
 
     @Test
